Create sticker_problem.py
This is an example python problem from interview questions. Here is my rough draft solution. 
basically takes each string and puts them in 2 dictionarys with the letter as the key and the value as the number of letters. Then it iterates through the control string's dict checking each character vs the input string's dict and if the difference between the two is more than the variable how_many_stickers, then it becomes the new how_many_stickers variable value. 

Once done with the iteration, then it prints the result of how many stickers need to be bought, thus solving the problem. 
I particularly like this solution because I made a function that doesn't hardcode a dictionary and you could do anything else as the control string quite easily if you needed to change it to a different test case. That same function is the one that is used to put the input string into a dict.