Create time_conversion.py
I could have done this easier if I had not broken up the string, but if we were doing this in the real world I expect I would want to break up the minutes and seconds as well. I'd also like to try using the time module to see if I can convert the string directly that way. I also want to do get better at knowing off the top of my head how to do fancier text formatting using {} for each var, but will do that later