Create lonely_integer.py
Started out by creating a dictionary from the list and then looping through the dictionary values again till I found 1, however I don't love that solution for a real world code. It would be more practical to have the # of values as the key in the dictionary just in case we actually did need to find multiple. 
Both solutions use 2 for loops, but since they are not nested the big O is still linear at O(n)