Rename counting_sort to counting_sort.py
Completed this one in 14 minutes. 
As a time saving strategy / code running strategy I like my method of making the code run and printing the test case. 
It "begins with the end in mind" and also gets me live code that I can tinker with while I get it  working.